+++ Essential_Feed/Essential_Feed/Feed Presentation/FeedPresenter.swift
@@ -1,0 +1,68 @@
+import Foundation
+
+public protocol FeedLoadingView {
+    func display(viewModel: FeedLoadingViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: FeedErrorViewModel {
+        return FeedErrorViewModel(message: nil)
+    }
+    
+    static func error(message: String) -> FeedErrorViewModel {
+        return FeedErrorViewModel(message: message)
+    }
+}
+
+public protocol FeedErrorView {
+    func display(viewModel: FeedErrorViewModel)
+}
+
+public protocol FeedView {
+    func display(viewModel: FeedViewModel)
+}
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public final class FeedPresenter {
+    private let errorView: FeedErrorView
+    private let loadingView: FeedLoadingView
+    private let feedView: FeedView
+    
+    public init(errorView: FeedErrorView, loadingView: FeedLoadingView, feedView: FeedView) {
+        self.errorView = errorView
+        self.loadingView = loadingView
+        self.feedView = feedView
+    }
+    
+    public static var title: String {
+        return NSLocalizedString("FEED_VIEW_TITLE", tableName: "Feed", bundle: Bundle(for: FeedPresenter.self), comment: "title for the feed view")
+    }
+    
+    private var feedLoadError: String {
+        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR", tableName: "Feed", bundle: Bundle(for: FeedPresenter.self), comment: "Error message displayed when we can't load the image feed from the server")
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(viewModel: .noError)
+        loadingView.display(viewModel: .init(isLoading: true))
+    }
+    
+    public func didFinishLoadingFeed(with feed: [FeedImage]) {
+        feedView.display(viewModel: .init(feed: feed))
+        loadingView.display(viewModel: .init(isLoading: false))
+    }
+    
+    public func didFinishLoadingFeed(with error: Error) {
+        errorView.display(viewModel: .error(message: feedLoadError))
+        loadingView.display(viewModel: .init(isLoading: false))
+    }
+}
